Fix up the 01 adaptive dialog sample
