+++ samples/csharp_dotnetcore/adaptive-dialog/01.multi-turn-prompt/Dialogs/RootDialog.lg
@@ -36,7 +36,7 @@
 # ConfirmPrompt
 - Is this ok?
 
-> This template uses inline expressions. Expressions are defined using adaptive expressions. 
+> This template uses in-line expressions. Expressions are defined using adaptive expressions. 
 > See https://aka.ms/adaptive-expressions to learn more.
 # FinalUserProfileReadOut
 - IF: ${turn.finalConfirmation == true}
@@ -44,7 +44,7 @@
 - ELSE:
     - Thanks. Your profile will not be kept.
 
-> Structured template defintion.
+> Structured template definition.
 # ModeOfTransportPrompt
 [Activity
     Text = ${ModeOfTransportPrompt.Text()}
@@ -52,5 +52,5 @@
 ]
 
 # ModeOfTransportPrompt.Text
-- What is your favoriate mode of transportion?
+- What is your favorite mode of transportation?
 - Please enter your mode of transport.
