+++ samples/csharp_dotnetcore/adaptive-dialog/01.multi-turn-prompt/wwwroot/default.htm
@@ -392,7 +392,7 @@
     <header class="header">
         <div class="header-inner-container">
             <div class="header-icon" style="display: inline-block"></div>
-            <div class="header-text" style="display: inline-block">Core Bot Sample</div>
+            <div class="header-text" style="display: inline-block">Multi-turn prompt sample</div>
         </div>
     </header>
     <div class="row">
